allow custom I/O to be used for credprompt
