+++ gwcli/connection/credprompt/credprompt_test.go
@@ -1,221 +1,59 @@
-//go:build ci
-
 /*************************************************************************
- * Copyright 2025 Gravwell, Inc. All rights reserved.
+ * Copyright 2026 Gravwell, Inc. All rights reserved.
  * Contact: <legal@gravwell.io>
  *
  * This software may be modified and distributed under the terms of the
  * BSD 2-clause license. See the LICENSE file for details.
  **************************************************************************/
 
-package credprompt
+package credprompt_test
 
 import (
+	"bytes"
 	"os"
+	"path/filepath"
+	"strings"
 	"testing"
-	"time"
 
-	tea "github.com/charmbracelet/bubbletea"
-	"github.com/charmbracelet/x/exp/teatest"
-	"github.com/gravwell/gravwell/v4/gwcli/internal/testsupport"
-	"github.com/gravwell/gravwell/v4/gwcli/stylesheet/hotkeys"
+	"github.com/Pallinder/go-randomdata"
+	"github.com/gravwell/gravwell/v4/gwcli/clilog"
+	"github.com/gravwell/gravwell/v4/gwcli/connection/credprompt"
+	"github.com/stretchr/testify/assert"
 )
 
-type output struct {
-	user         string
-	pass         string
-	killed       bool
-	userSelected bool
-}
+// Tests the public, exposed Collect api and that I/O can be redirected.
+func TestCollect(t *testing.T) {
+	var (
+		usernamePrefill = "half"      // what text is prefilled in the username field via parameter
+		usernameEntered = "-username" // what text will be entered from in
+		password        = "mypass"
+		logpath         = filepath.Join(t.TempDir(), randomdata.Alphanumeric(10)+".txt")
+	)
 
-// NOTE: this testing package relies on teatest, which is an experimental package at the time of authorship (~June 2025).
+	// clilog needs to be spinning
+	clilog.InitializeFromArgs([]string{"--" + clilog.FlagLogPath.Name + "=" + logpath})
 
-// TestCredPrompt_TeaTest runs interactivity tests against the cred prompt model
-func TestCredPrompt_TeaTest(t *testing.T) {
-	tests := []struct {
-		name             string
-		input            func(tm *teatest.TestModel, expected output) // used to send messages to the model, to mimic user input
-		expected         output                                       // what the final model's values should look like
-		timeoutDur       time.Duration                                // stop waiting on the final model after this much time
-		expectingTimeout bool                                         // are we expecting this test to timeout while awaiting the final model
-	}{
-		{"normal u/p", func(tm *teatest.TestModel, expected output) {
-			tm.Type(expected.user)
-			testsupport.TTSendSpecial(tm, testsupport.SendHotkey(hotkeys.Invoke).Type)
-			tm.Type(expected.pass)
-			testsupport.TTSendSpecial(tm, testsupport.SendHotkey(hotkeys.Invoke).Type)
-		}, output{"Blitzo", "TheOIsSilent", false, false}, 2 * time.Second, false},
-		{"garbage after submitting", func(tm *teatest.TestModel, expected output) {
-			tm.Type(expected.user)
-			testsupport.TTSendSpecial(tm, testsupport.SendHotkey(hotkeys.Invoke).Type)
-			tm.Type(expected.pass)
-			testsupport.TTSendSpecial(tm, testsupport.SendHotkey(hotkeys.Invoke).Type)
-
-			// this should not be captured by the prompt
-			tm.Type("should not be caught")
-			tm.Send(tea.KeyMsg(tea.Key{Type: tea.KeyCtrlC, Runes: []rune{rune(tea.KeyCtrlC)}}))
-		}, output{"Moxxie", "Milly", false, false}, 2 * time.Second, false},
-		{"global kill key", func(tm *teatest.TestModel, expected output) {
-			tm.Type(expected.user)
-			testsupport.TTSendSpecial(tm, testsupport.SendHotkey(hotkeys.CursorDown).Type)
-			tm.Type(expected.pass)
-
-			// kill with a sigint
-			testsupport.TTSendSpecial(tm, tea.KeyCtrlC)
-			// this should not be captured by the prompt
-			tm.Type("should not be caught")
-		}, output{"Stolas", "Blitzy", true, false}, 2 * time.Second, false},
-		{"child kill key", func(tm *teatest.TestModel, expected output) {
-			tm.Type(expected.user)
-
-			// kill with a sigint
-			testsupport.TTSendSpecial(tm, tea.KeyEsc)
-
-			// this should not be captured by the prompt
-			tm.Type("should not be caught")
-		}, output{"Loona", "", true, true}, 2 * time.Second, false},
-		{"wrap", func(tm *teatest.TestModel, expected output) {
-			tm.Type(expected.user)
-			testsupport.TTSendSpecial(tm, testsupport.SendHotkey(hotkeys.CursorDown).Type)
-			tm.Type(expected.pass)
-			testsupport.TTSendSpecial(tm, testsupport.SendHotkey(hotkeys.CursorDown).Type)
-			testsupport.TTSendSpecial(tm, testsupport.SendHotkey(hotkeys.CursorUp).Type)
-			testsupport.TTSendSpecial(tm, testsupport.SendHotkey(hotkeys.Invoke).Type)
-		}, output{"Fizzarolli", "Oops", false, false}, 2 * time.Second, false},
-		{"timeout", func(tm *teatest.TestModel, expected output) {
-			testsupport.TTSendSpecial(tm, testsupport.SendHotkey(hotkeys.Invoke).Type)
-			tm.Type("some password that should not get returned")
-			testsupport.TTSendSpecial(tm, testsupport.SendHotkey(hotkeys.CursorUp).Type)
-		}, output{}, 3 * time.Second, true},
+	// predefine IO
+	var (
+		in  = bytes.NewBufferString(usernameEntered + "\r" + password + "\r") // REMINDER: BubbleTea reads carriage returns as "enter"
+		out = strings.Builder{}
+	)
+	gotUser, gotPass, gotErr := credprompt.Collect(usernamePrefill, in, &out) // prefill
+	if gotErr != nil {
+		t.Fatal(gotErr)
 	}
+	assert.Equal(t, usernamePrefill+usernameEntered, gotUser)
+	assert.Equal(t, password, gotPass)
 
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			tm, ch := spawnModel(t)
+	if t.Failed() {
+		// let's see what is in the output buffer
+		t.Log("final output:\n", out.String())
 
-			// execute commands against the test model
-			tt.input(tm, tt.expected)
-
-			// check results
-			out, timedout := awaitFinal(t, ch, tt.timeoutDur)
-			compareFinal(t, out, tt.expected, timedout, tt.expectingTimeout)
-		})
-	}
-
-}
-
-// Test_collect tests .Collect() via the internal subroutine that .Collect() calls under the hood.
-// Does not actually rely on teatest; instead, passes a tea program with a mocked input and interacts via external .Send()ing.
-func Test_collect(t *testing.T) {
-
-	tests := []struct {
-		name         string
-		input        func(prog *tea.Program)
-		expectedUser string
-		expectedPass string
-		expectedErr  bool
-	}{
-		{"normal u/p", func(prog *tea.Program) {
-			prog.Send(tea.KeyMsg(tea.Key{Type: tea.KeyRunes, Runes: []rune{'u'}}))
-			testsupport.TTSendSpecial(prog, testsupport.SendHotkey(hotkeys.Invoke).Type)
-			testsupport.TTSendSpecial(prog, testsupport.SendHotkey(hotkeys.Invoke).Type)
-
-		}, "u", "", false},
-		{"killed", func(prog *tea.Program) {
-			testsupport.TTSendSpecial(prog, tea.KeyCtrlC)
-		}, "", "", true},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			result := make(chan struct {
-				user string
-				pass string
-				err  error
-			})
-
-			// spawn a model
-			m := New("")
-			read, _, err := os.Pipe()
-			if err != nil {
-				t.Fatal(err)
-			}
-			// make the model read out of an open pipe
-			prog := tea.NewProgram(m, tea.WithInput(read))
-
-			// spin off the actual TUI via Collect()
-			go func() {
-				u, p, err := collect("", prog)
-
-				result <- struct {
-					user string
-					pass string
-					err  error
-				}{u, p, err}
-			}()
-
-			// send in mock-user input
-			tt.input(prog)
-
-			// await results
-			r := <-result
-			if (r.err != nil) != tt.expectedErr {
-				t.Errorf("Expected error? %v. Actual error: %v", tt.expectedErr, r.err)
-			} else if r.user != tt.expectedUser {
-				t.Error("Unexpected user:", testsupport.ExpectedActual(tt.expectedErr, r.err))
-			} else if r.pass != tt.expectedPass {
-				t.Error("Unexpected password:", testsupport.ExpectedActual(tt.expectedErr, r.err))
-			}
-		})
-	}
-
-}
-
-// spawnModel spins off a credprompt returns a channel that can be read from after the model exists to get its final state.
-func spawnModel(t *testing.T) (*teatest.TestModel, chan tea.Model) {
-	t.Helper()
-	// create a channel for us to receive the final model on
-	result := make(chan tea.Model)
-
-	// spawn a model
-	m := New("")
-	tm := teatest.NewTestModel(t, m, teatest.WithInitialTermSize(300, 100))
-	go func() {
-		final := tm.FinalModel(t, teatest.WithFinalTimeout(10*time.Second))
-		result <- final
-		close(result)
-	}()
-
-	return tm, result
-}
-
-// awaitFinal pulls data from the final struct and returns it for easy evaluation.
-//
-// Returns the final values contained in the credmodel or that we timed-out waiting for it.
-func awaitFinal(t *testing.T, ch <-chan tea.Model, timeout time.Duration) (output, bool) {
-	t.Helper()
-
-	select {
-	case final := <-ch:
-		cm, ok := final.(credModel)
-		if !ok {
-			t.Fatal("failed to assert final model to a credModel")
+		b, err := os.ReadFile(logpath)
+		if err != nil {
+			t.Log("failed to dump log file:", err)
+		} else {
+			t.Log(string(b))
 		}
-		// check the results
-		return output{cm.UserTI.Value(), cm.PassTI.Value(), cm.killed, cm.userSelected}, false
-
-	case <-time.After(timeout):
-		return output{}, true
-	}
-}
-
-// compareFinal tests the actual output of awaitFinal against the expected output of awaitFinal, erroring on deltas.
-func compareFinal(t *testing.T, actual, expected output, timedOut, expectedTimedOut bool) {
-	t.Helper()
-
-	if timedOut != expectedTimedOut {
-		t.Fatal("timed out waiting for credModel to finish")
-	} else if actual != expected {
-		t.Error("incorrect final state:", testsupport.ExpectedActual(expected, actual))
 	}
 }
