+++ gwcli/tree/ingest/ingest_noci_test.go
@@ -46,7 +46,7 @@ func Test_autoingest(t *testing.T) {
 		t.Fatal(err)
 	} else if err := connection.Initialize(server, false, true, path.Join(dir, "dev.log")); err != nil {
 		t.Fatal(err)
-	} else if err := connection.Login(username, &password, nil, true); err != nil {
+	} else if err := connection.Login(username, &password, nil, true, os.Stdin, os.Stdout); err != nil {
 		t.Fatal(err)
 	}
 
@@ -277,7 +277,7 @@ func TestNewIngestActionRun(t *testing.T) {
 		t.Fatal(err)
 	} else if err := connection.Initialize(server, false, true, path.Join(t.TempDir(), "dev.log")); err != nil {
 		t.Fatal(err)
-	} else if err := connection.Login(username, &password, nil, true); err != nil {
+	} else if err := connection.Login(username, &password, nil, true, os.Stdin, os.Stdout); err != nil {
 		t.Fatal(err)
 	}
 	tDir := t.TempDir()
