+++ gwcli/utilities/querysupport/querysupport_noci_test.go
@@ -39,7 +39,7 @@ func Test_GetResultsForDataScope(t *testing.T) {
 	if err := connection.Initialize(server, false, true, ""); err != nil {
 		panic(err)
 	}
-	if err := connection.Login(user, &password, nil, true); err != nil {
+	if err := connection.Login(user, &password, nil, true, nil, nil); err != nil {
 		panic(err)
 	}
